# [Tooling] Enable U Test by default, add tooltip about repetition count.